Refactor SparkSession to fix Spark-4.0 build
