+++ sql-plugin/src/main/scala/org/apache/spark/sql/rapids/execution/InternalColumnarRddConverter.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2019-2024, NVIDIA CORPORATION. All rights reserved.
+ * Copyright (c) 2019-2025, NVIDIA CORPORATION. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -661,11 +661,11 @@ object InternalColumnarRddConverter extends Logging {
         s"not currently supported data types for columnar.")
     }
     //This config lets the plugin tag the columnar transition we care about so we know what we got.
-    df.sqlContext.setConf(RapidsConf.EXPORT_COLUMNAR_RDD.key, "true")
+    df.sparkSession.sqlContext.setConf(RapidsConf.EXPORT_COLUMNAR_RDD.key, "true")
     val input = try {
       df.rdd
     } finally {
-      df.sqlContext.setConf(RapidsConf.EXPORT_COLUMNAR_RDD.key, "false")
+      df.sparkSession.sqlContext.setConf(RapidsConf.EXPORT_COLUMNAR_RDD.key, "false")
     }
     var batch: Option[RDD[ColumnarBatch]] = None
     // If we are exporting the data we will see
@@ -715,7 +715,7 @@ object InternalColumnarRddConverter extends Logging {
     val b = batch.getOrElse({
       // We have to fall back to doing a slow transition.
       val converters = new GpuExternalRowToColumnConverter(schema)
-      val conf = new RapidsConf(df.sqlContext.sparkSession.sessionState.conf)
+      val conf = new RapidsConf(df.sparkSession.sessionState.conf)
       val goal = TargetSize(conf.gpuTargetBatchSizeBytes)
       input.mapPartitions { rowIter =>
         new ExternalRowToColumnarIterator(rowIter, schema, goal, converters)
